Adding update function
