+++ backend/module/product_module/route/index.js
@@ -1,0 +1,9 @@
+import express from "express";
+import authenticate from "../../../middleware/jwtAuthenticationMiddleware.js";
+import controller from "../controller/productController.js";
+
+const route = express.Router();
+
+route.post("/register_product", authenticate, controller.productRegister);
+
+export default route;
